requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==2.0.36
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
flasgger
# Database
SQLAlchemy==2.0.36
alembic==1.14.0

# Validation and serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Environment management
python-dotenv==1.0.0

# WSGI server
gunicorn==21.2.0

# Database drivers
PyMySQL==1.1.0          # MySQL (optional)

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
flake8==6.1.0
black==23.12.1